+++ tools/tetl/etl.go
@@ -309,7 +309,7 @@ func InitSpec(t *testing.T, bp api.BaseParams, etlName, comm string) (xid string
 	msg := &etl.InitSpecMsg{}
 	msg.EtlName = etlName
 	msg.CommTypeX = comm
-	msg.Timeout = cos.Duration(time.Minute * 2) // manually increase timeout in testing environment
+	msg.InitTimeout = cos.Duration(time.Minute * 2) // manually increase timeout in testing environment
 	spec, err := GetTransformYaml(etlName)
 	tassert.CheckFatal(t, err)
 	msg.Spec = spec
@@ -421,7 +421,7 @@ func SpecToInitMsg(spec []byte /*yaml*/) (*etl.InitSpecMsg, error) {
 	}
 	// Check annotations.
 	msg.CommTypeX = podTransformCommType(pod)
-	if msg.Timeout, err = podTransformTimeout(errCtx, pod); err != nil {
+	if msg.InitTimeout, err = podTransformTimeout(errCtx, pod); err != nil {
 		return msg, err
 	}